Crate c3po

Structs

Config

Conn
A smart wrapper around a connection which stores it back in the pool when it is dropped.

Pool
An asynchronous, single-threaded connection pool.

Type Definitions

ConnFuture
Future yielded by Pool::connection. Optimized not to allocate when pulling an idle future out of the pool.